refactor: remove all remaining enums from ContractCallServicePrecompile
